// src/types.ts
  export interface Matches {
    matches: Match[];
    cursor: string;
  }

  export interface Match {
    id: string;
    start_time: string;
    end_time: string;
    game_duration: number;
    game_region: any;
    region: string;
    winning_team: string;
    game_mode: string;
    players: MatchPlayer[];
  }

  export interface MatchPlayer {
    id: string;
    display_name: string;
    is_ranked: boolean;
    mmr: number;
    mmr_change: number;
    rank: string;
    rank_image: string;
    flags: Flag[];
    team: string;
    hero_id: number;
    role: any;
    minions_killed: number;
    lane_minions_killed: number;
    neutral_minions_killed: number;
    neutral_minions_team_jungle: number;
    neutral_minions_enemy_jungle: number;
    kills: number;
    deaths: number;
    assists: number;
    largest_killing_spree: number;
    largest_multi_kill: number;
    total_damage_dealt: number;
    physical_damage_dealt: number;
    magical_damage_dealt: number;
    true_damage_dealt: number;
    largest_critical_strike: any;
    total_damage_dealt_to_heroes: number;
    physical_damage_dealt_to_heroes: number;
    magical_damage_dealt_to_heroes: number;
    true_damage_dealt_to_heroes: number;
    total_damage_dealt_to_structures: number;
    total_damage_dealt_to_objectives: number;
    total_damage_taken: number;
    physical_damage_taken: number;
    magical_damage_taken: number;
    true_damage_taken: number;
    total_damage_taken_from_heroes: number;
    physical_damage_taken_from_heroes: number;
    magical_damage_taken_from_heroes: number;
    true_damage_taken_from_heroes: number;
    total_damage_mitigated: number;
    total_healing_done: number;
    item_healing_done: any;
    crest_healing_done: any;
    utility_healing_done: any;
    total_shielding_received: any;
    wards_placed: number;
    wards_destroyed: number;
    gold_earned: number;
    gold_spent: number;
    inventory_data: any;
    performance_score: number;
    performance_title: string;
    level: number;
  }

  export interface Flag {
    identifier: string;
    text: string;
    color: string;
  }